requirements.txt
# Core dependencies
Flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
python-dotenv==1.0.0

# LLM providers
# The system supports multiple LLM providers. Install the ones you need:
google-genai==0.6.0
groq==0.4.2

# Optional LLM providers (uncomment and install as needed)
# openai==1.3.0
# anthropic==0.7.0

# Development dependencies (optional)
# pytest==7.4.3
# black==23.12.0
